Address remaining PR review comments
