+++ include/rut/runtime/epoll_backend.h
@@ -104,17 +104,4 @@ private:
     static void decode_data(u64 data, u32& conn_id, IoEventType& type);
 };
 
-#if defined(RUT_TESTING)
-// Test-only seam for forcing specific TLS state-machine transitions in epoll.
-struct EpollTlsHooks {
-    i32 (*ssl_accept)(SSL* ssl);
-    i32 (*ssl_read)(SSL* ssl, void* buf, i32 len);
-    i32 (*ssl_write)(SSL* ssl, const void* buf, i32 len);
-    i32 (*ssl_get_error)(SSL* ssl, i32 rc);
-};
-
-void set_epoll_tls_hooks_for_test(const EpollTlsHooks* hooks);
-void reset_epoll_tls_hooks_for_test();
-#endif
-
 }  // namespace rut
