+++ src/runtime/epoll_backend.cc
@@ -1,6 +1,7 @@
 #include "rut/runtime/epoll_backend.h"
 
 #include "core/expected.h"
+#include "epoll_tls_test_hooks.h"
 #include "rut/runtime/error.h"
 
 #include <errno.h>
@@ -18,15 +19,6 @@ static constexpr u32 kListenConnId = 0xFFFFFF;
 static constexpr u32 kTimerConnId = 0xFFFFFE;
 
 namespace {
-
-#if !defined(RUT_TESTING)
-struct EpollTlsHooks {
-    i32 (*ssl_accept)(SSL* ssl);
-    i32 (*ssl_read)(SSL* ssl, void* buf, i32 len);
-    i32 (*ssl_write)(SSL* ssl, const void* buf, i32 len);
-    i32 (*ssl_get_error)(SSL* ssl, i32 rc);
-};
-#endif
 
 i32 default_ssl_accept(SSL* ssl) {
     return SSL_accept(ssl);
@@ -47,7 +39,6 @@ const EpollTlsHooks* active_tls_hooks = &default_tls_hooks;
 
 }  // namespace
 
-#if defined(RUT_TESTING)
 void set_epoll_tls_hooks_for_test(const EpollTlsHooks* hooks) {
     active_tls_hooks = hooks ? hooks : &default_tls_hooks;
 }
@@ -55,7 +46,6 @@ void set_epoll_tls_hooks_for_test(const EpollTlsHooks* hooks) {
 void reset_epoll_tls_hooks_for_test() {
     active_tls_hooks = &default_tls_hooks;
 }
-#endif
 
 static i32 map_tls_error(SSL* ssl, i32 rc) {
     i32 err = active_tls_hooks->ssl_get_error(ssl, rc);
@@ -549,15 +539,8 @@ u32 EpollBackend::wait(IoEvent* events, u32 max_events, Connection* conns, u32 m
             if (conn_id >= kMaxFdMap || conn_id >= max_conns) continue;
             auto& ss = send_state[conn_id];
             if (ss.remaining == 0 || !ss.src || ss.fd < 0) {
-                i32 rfd = -1;
-                IoEventType recv_type = IoEventType::Recv;
-                if (upstream_fd_map[conn_id] >= 0 && upstream_fd_map[conn_id] == ss.fd) {
-                    rfd = upstream_fd_map[conn_id];
-                    recv_type = IoEventType::UpstreamRecv;
-                } else {
-                    rfd = downstream_fd_map[conn_id];
-                }
-                if (rfd >= 0) set_fd_interest(epoll_fd, rfd, conn_id, recv_type, EPOLLIN);
+                // No outstanding send associated with this connection; ignore
+                // spurious EPOLLOUT rather than rearming the wrong fd.
                 continue;
             }
